requirements.txt
allure-pytest==2.8.18
allure-python-commons==2.8.18
apipkg==1.5
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
coverage==5.2.1
execnet==1.7.1
idna==2.10
iniconfig==1.0.1
more-itertools==8.4.0
mysql-connector==2.2.9
packaging==20.4
pep8==1.7.1
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pycryptodome==3.9.8
pyfiglet==0.8.post1
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.0.1
pytest-assume==2.3.2
pytest-cache==1.0
pytest-cov==2.10.1
pytest-dependency==0.5.1
pytest-flakes==4.0.1
pytest-html==2.1.1
pytest-metadata==1.10.0
pytest-ordering==0.6
pytest-pep8==1.0.6
pytest-rerunfailures==9.0
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
